Move two new routines from stream/base.py to stream/makeNotation.py.
